requirements.txt
pyserial
pySerialTransfer
numpy
dearpygui